Rename modules and move mldata classes.
