+++ datarobot_provider/_experimental/deprecated_example_dags/datarobot_create_project_from_ai_catalog_dag.py
@@ -22,7 +22,7 @@ from datetime import datetime
 
 from airflow.decorators import dag
 
-from datarobot_provider.operators.ai_catalog import UploadDatasetOperator
+from datarobot_provider.operators.data_registry import UploadDatasetOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
