+++ datarobot_provider/_experimental/deprecated_example_dags/datarobot_create_project_from_dataset_version_dag.py
@@ -23,7 +23,7 @@ from datetime import datetime
 
 from airflow.decorators import dag
 
-from datarobot_provider.operators.ai_catalog import UpdateDatasetFromFileOperator
+from datarobot_provider.operators.data_registry import UpdateDatasetFromFileOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
