+++ datarobot_provider/_experimental/deprecated_example_dags/datarobot_datetime_partitioning_pipeline_dag.py
@@ -12,7 +12,7 @@ from datarobot import AUTOPILOT_MODE
 from datarobot.enums import DATETIME_AUTOPILOT_DATA_SELECTION_METHOD
 from datarobot.helpers.partitioning_methods import construct_duration_string
 
-from datarobot_provider.operators.ai_catalog import UploadDatasetOperator
+from datarobot_provider.operators.data_registry import UploadDatasetOperator
 from datarobot_provider.operators.autopilot import StartAutopilotOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
 from datarobot_provider.sensors.datarobot import AutopilotCompleteSensor
