+++ datarobot_provider/example_dags/dataprep_wrangler_autopilot.py
@@ -9,9 +9,9 @@
 import datarobot as dr
 from airflow.decorators import dag
 
-from datarobot_provider.operators.ai_catalog import CreateDatasetFromRecipeOperator
-from datarobot_provider.operators.ai_catalog import CreateWranglingRecipeOperator
-from datarobot_provider.operators.ai_catalog import UploadDatasetOperator
+from datarobot_provider.operators.data_registry import CreateDatasetFromRecipeOperator
+from datarobot_provider.operators.data_prep import CreateWranglingRecipeOperator
+from datarobot_provider.operators.data_registry import UploadDatasetOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
 from datarobot_provider.operators.datarobot import GetOrCreateUseCaseOperator
 from datarobot_provider.operators.datarobot import SelectBestModelOperator
