+++ datarobot_provider/example_dags/featrue_discovery_retraining_and_scoring.py
@@ -9,13 +9,13 @@
 from airflow.decorators import dag
 from datarobot import AUTOPILOT_MODE
 
-from datarobot_provider.operators.ai_catalog import UploadDatasetOperator
+from datarobot_provider.operators.data_registry import UploadDatasetOperator
 from datarobot_provider.operators.autopilot import StartAutopilotOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
 from datarobot_provider.operators.datarobot import GetOrCreateUseCaseOperator
 from datarobot_provider.operators.datarobot import SelectBestModelOperator
 from datarobot_provider.operators.deployment import ReplaceModelOperator
-from datarobot_provider.operators.feature_discovery import CreateFeatureDiscoveryRecipeOperator
+from datarobot_provider.operators.data_prep import CreateFeatureDiscoveryRecipeOperator
 from datarobot_provider.operators.model_registry import CreateRegisteredModelVersionOperator
 from datarobot_provider.sensors.datarobot import AutopilotCompleteSensor
