+++ datarobot_provider/example_dags/model_training_xgboost.py
@@ -9,10 +9,10 @@
 import datarobot as dr
 from airflow.decorators import dag
 
-from datarobot_provider.operators.ai_catalog import CreateDatasetFromRecipeOperator
-from datarobot_provider.operators.ai_catalog import CreateOrUpdateDataSourceOperator
-from datarobot_provider.operators.ai_catalog import CreateWranglingRecipeOperator
-from datarobot_provider.operators.connections import GetDataStoreOperator
+from datarobot_provider.operators.data_registry import CreateDatasetFromRecipeOperator
+from datarobot_provider.operators.data_registry import CreateOrUpdateDataSourceOperator
+from datarobot_provider.operators.data_prep import CreateWranglingRecipeOperator
+from datarobot_provider.operators.data_registry import GetDataStoreOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
 from datarobot_provider.operators.datarobot import GetOrCreateUseCaseOperator
 from datarobot_provider.operators.datarobot import GetProjectBlueprintsOperator
